refactore settings
