+++ config/local.toml.bak
@@ -1,0 +1,92 @@
+####################################################################################
+#                                                                                  #
+# Kellnr configuration.                                                            #
+# For more details, see: https://www.bitfalter.com/documentation#Configuration     #
+#                                                                                  #
+####################################################################################
+
+
+
+####################################################################################
+#                                                                                  #
+# Values used only on the first startup. Can be changed using the UI on runtime.   #
+#                                                                                  #
+####################################################################################
+admin_pwd = "admin"
+admin_token = "Zy9HhJ02RJmg0GCrgLfaCVfU6IwDfhXD"
+
+
+####################################################################################
+#                                                                                  #
+# Values used on each start of Kellnr. Overwrite and restart Kellnr to change.     #
+#                                                                                  #
+####################################################################################
+
+# Address where Kellnr will be reachable
+# E.g. https://kellnr.example.com:443
+api_address = "127.0.0.1"
+# The port where Kellnr starts listening for incoming connections
+api_port = 8000
+# If a proxy is used in front of Kellnr, the port of the proxy can be specified here
+# If no proxy is used, it is the same as the "api_port"
+api_port_proxy = 8000 
+# Either "https" or "http". Use in combination with a reverse proxy that provides HTTPS.
+api_protocol = "http"
+
+# Directory where Kellnr stores all its data, e.g. crates, incides etc.
+data_dir = "/Users/hausi/repos/kdata"
+
+# Seconds until a user is logged out automatically after inactivity in the UI
+session_age_seconds = 28800
+
+# Address where the API and web server is started. Usually no change is needed.
+web_address = "0.0.0.0"
+
+# Set to "true" to enable the crates.io proxy. The the official Kellnr documentation
+# for more information.
+crates_io_proxy = true
+
+# Number of threads used to keep the crates.io proxy up to date.
+# A too high number can lead to exhausting the available database connection.
+crates_io_num_threads = 20
+
+# Set the log level to "trace", "debug", "info", "warn", or "error".
+log_level = "trace"
+
+# Set the log level for the underlaying web framework to "trace", "debug", "info", "warn", or "error".
+log_level_web_server = "debug"
+
+# Set the log format to "compact", "pretty" or "json".
+log_format = "compact"
+
+# Enable or disable automatic rustdoc generation for uploaded crates
+rustdoc_auto_gen = true
+
+# Number of elements of the in-memory cache to take load of the disk for crate downloads
+# If set to "0", the cache is disabled.
+cache_size = 1000
+
+# Max size of a crate that can be uploaded to Kellnr in MB
+max_crate_size = 1000
+
+# Max size of a crate docs that can be uploaded to Kellnr in MB
+max_docs_size = 1000
+
+####################################################################################
+# Kellnr Enterprise only                                                           #
+####################################################################################
+
+# Enable required authentication for crate pulls.
+# If set to "false", anyone can download crates from Kellnr. Upload always requires authentication.
+# This feature is currently a nightly only feature in cargo and might break in the future,
+# as the RFC is not fully done yet.
+auth_required = false
+
+# Configure Postgresql as the database backend instead of Sqlite
+[postgresql]
+enabled = true
+address = "192.168.178.2"
+port = 5432
+db = "kellnr-meta-db"
+user = "admin"
+pwd = "admin"
