+++ crates/settings/src/docs.rs
@@ -1,0 +1,13 @@
+use serde::{Serialize, Deserialize};
+
+#[derive(Debug, Deserialize, Serialize, Eq, PartialEq, Clone)]
+pub struct Docs {
+    pub enabled: bool,
+    pub max_size: u64,
+}
+
+impl Default for Docs {
+    fn default() -> Self {
+        Self { enabled: false, max_size: 100*1000 }
+    }
+}
