+++ crates/settings/src/lib.rs
@@ -1,191 +1,15 @@
-use config::{Config, ConfigError, Environment, File};
-use deserialize_with::DeserializeWith;
-use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
-use std::str::FromStr;
-use std::{
-    convert::TryFrom,
-    env,
-    net::IpAddr,
-    path::{self, Path},
-};
-
+pub mod docs;
+pub mod local;
+pub mod log;
+pub mod origin;
+pub mod postgresql;
+pub mod protocol;
+pub mod proxy;
+pub mod registry;
+pub mod settings;
+pub mod startup;
 pub mod constants;
 mod deserialize_with;
-mod log_format;
-mod log_level;
-mod protocol;
 
-pub use log_format::LogFormat;
-pub use protocol::Protocol;
-
-#[derive(Debug, Deserialize, Serialize, Eq, PartialEq, Default, Clone)]
-pub struct Postgresql {
-    pub enabled: bool,
-    pub address: String,
-    pub port: u16,
-    pub db: String,
-    pub user: String,
-    #[serde(skip_serializing, default)]
-    pub pwd: String,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct Settings {
-    pub admin_pwd: String,
-    pub data_dir: String,
-    pub session_age_seconds: u64,
-    pub api_address: String,
-    pub api_port: u16,
-    pub api_port_proxy: u16,
-    #[serde(
-        deserialize_with = "Protocol::deserialize_with",
-        default = "Protocol::default"
-    )]
-    pub api_protocol: Protocol,
-    pub web_address: IpAddr,
-    pub admin_token: String,
-    #[serde(default)]
-    pub crates_io_proxy: bool,
-    pub crates_io_num_threads: usize,
-    #[serde(deserialize_with = "tracing::Level::deserialize_with")]
-    pub log_level: tracing::Level,
-    #[serde(deserialize_with = "tracing::Level::deserialize_with")]
-    pub log_level_web_server: tracing::Level,
-    #[serde(deserialize_with = "LogFormat::deserialize_with")]
-    pub log_format: LogFormat,
-    #[serde(default)]
-    pub postgresql: Postgresql,
-    pub rustdoc_auto_gen: bool,
-    #[serde(default)]
-    pub cache_size: u64,
-    pub max_crate_size: usize,
-    pub max_docs_size: usize,
-    #[serde(default)]
-    pub auth_required: bool,
-}
-
-impl TryFrom<&Path> for Settings {
-    type Error = ConfigError;
-
-    fn try_from(config_path: &Path) -> Result<Self, Self::Error> {
-        let default_file = Settings::join_path(config_path, "default")?;
-        let env = env::var("RUN_MODE").unwrap_or_else(|_| "development".into());
-        let env_file = Settings::join_path(config_path, &env)?;
-        let local_file = Settings::join_path(config_path, "local")?;
-
-        let s = Config::builder()
-            // Set default values of settings that where added after the first Kellnr release
-            // and thus can be missing on the customer site.
-            .set_default("log_level", "info")?
-            .set_default("log_level_web_server", "warn")?
-            .set_default("log_format", "compact")?
-            .set_default("api_port_proxy", 8000)?
-            .set_default("rustdoc_auto_gen", true)?
-            .set_default("max_crate_size", 100 * 1000)?
-            .set_default("max_docs_size", 100 * 1000)?
-            .set_default("crates_io_num_threads", 10)?
-            // Start off by merging in the "default" configuration file
-            .add_source(File::with_name(&default_file))
-            // Add in the current environment file
-            // Default to 'development' env
-            // Note that this file is _optional_
-            .add_source(File::with_name(&env_file).required(false))
-            // Add in a local configuration file
-            // This file shouldn't be checked in to git
-            .add_source(File::with_name(&local_file).required(false))
-            // Add in settings from the environment (with a prefix of KELLNR)
-            // Eg. `KELLNR_DEBUG=1 ./target/app` would set the `debug` key
-            .add_source(
-                Environment::with_prefix("KELLNR")
-                    .prefix_separator("_")
-                    .separator("__"),
-            )
-            .build()?;
-
-        // You can deserialize (and thus freeze) the entire configuration as
-        s.try_deserialize()
-    }
-}
-
-impl Settings {
-    fn join_path(config_path: &Path, file: &str) -> Result<String, ConfigError> {
-        config_path
-            .join(file)
-            .to_str()
-            .map(|x| x.to_string())
-            .ok_or_else(|| ConfigError::Message("Invalid UTF-8 string".to_string()))
-    }
-
-    pub fn bin_path(&self) -> path::PathBuf {
-        path::PathBuf::from(&self.data_dir).join("crates")
-    }
-
-    pub fn doc_queue_path(&self) -> PathBuf {
-        path::PathBuf::from(&self.data_dir).join("doc_queue")
-    }
-
-    pub fn sqlite_path(&self) -> path::PathBuf {
-        path::PathBuf::from(&self.data_dir).join("db.sqlite")
-    }
-
-    pub fn docs_path(&self) -> path::PathBuf {
-        path::PathBuf::from(&self.data_dir).join("docs")
-    }
-
-    pub fn base_path(&self) -> path::PathBuf {
-        path::PathBuf::from(&self.data_dir).join("git")
-    }
-
-    pub fn crates_io_bin_path(&self) -> path::PathBuf {
-        path::PathBuf::from(&self.data_dir).join("cratesio")
-    }
-}
-
-impl Default for Settings {
-    fn default() -> Self {
-        Self {
-            admin_pwd: "admin".to_string(),
-            data_dir: "/tmp/kellnr".to_string(), // Differs from default.toml for testing
-            session_age_seconds: 28800,
-            api_address: "127.0.0.1".to_string(),
-            api_port: 8000,
-            api_port_proxy: 8000,
-            api_protocol: Protocol::Http,
-            web_address: IpAddr::from_str("0.0.0.0").unwrap(),
-            admin_token: "Zy9HhJ02RJmg0GCrgLfaCVfU6IwDfhXD".to_string(),
-            crates_io_proxy: false,
-            crates_io_num_threads: 10,
-            log_level: tracing::Level::INFO,
-            log_level_web_server: tracing::Level::WARN,
-            log_format: LogFormat::Compact,
-            postgresql: Postgresql {
-                enabled: false,
-                address: "localhost".to_string(),
-                port: 5432,
-                db: "kellnr".to_string(),
-                user: "".to_string(),
-                pwd: "".to_string(),
-            },
-            rustdoc_auto_gen: false, // Differs from default.toml for testing
-            cache_size: 1000,
-            max_crate_size: 100 * 1000,
-            max_docs_size: 100 * 1000,
-            auth_required: false,
-        }
-    }
-}
-
-pub fn get_settings() -> Result<Settings, ConfigError> {
-    let path = if Path::new("./config").exists() {
-        Path::new("./config")
-    } else if Path::new("../config").exists() {
-        Path::new("../config")
-    } else if Path::new("../../config").exists() {
-        Path::new("../../config")
-    } else {
-        return Err(ConfigError::NotFound("Config folder not found".to_string()));
-    };
-
-    Settings::try_from(path)
-}
+pub use settings::Settings;
+pub use settings::get_settings;
