+++ crates/settings/src/proxy.rs
@@ -1,0 +1,13 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize, Eq, PartialEq, Clone)]
+pub struct Proxy {
+    pub enabled: bool,
+    pub num_threads: u64,
+}
+
+impl Default for Proxy {
+    fn default() -> Self {
+        Self { enabled: false, num_threads: 10 }
+    }
+}
